Switch to using Angular Universal
This is an initial working commit using Angular Universal to provide SSR through an express server. It is not ready for production.

A single persistent web socket connection is maintained by the server for use in fulfilling non-authenticated Expanse requests to build pages before being sent to the client. Clients will not re-request this information from the server.

As a first pass, things are working but should be cleaned up. Notably:
- Stress tests need to be run to find any issues with the server web socket at volume
- Stress tests need to be run to determine rendering load on the server
- When client pages initially load, some flashes of unstyled content show, including a large version of the logo
- Though the page HTML includes required content for an app/game, for instance, it is not yet a clean display with javascript disabled
- Title and meta tags should be updated (currently lives in a separate branch)
- The websocket incorrectly registers a second request from the client for each app listing page-load, resulting in an app search being performed
- There may be memory leaks